moving all error handling classes to dedicated package
