+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/error/BinderErrorConfigurer.java
@@ -14,8 +14,12 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.binder;
+package org.springframework.cloud.stream.error;
 
+
+import org.springframework.cloud.stream.binder.AbstractBinder;
+import org.springframework.cloud.stream.binder.Binding;
+import org.springframework.cloud.stream.binder.ConsumerProperties;
 
 /**
  * Binder implementations should implement their Error Handling strategy using sub classes of this interface.
@@ -25,7 +29,30 @@ package org.springframework.cloud.stream.binder;
  * @since 1.3
  */
 public interface BinderErrorConfigurer<C extends ConsumerProperties, T> {
+	/**
+	 * Registers any error infrastructure needed by the binder to handle errors.
+	 *
+	 * @param binding target binding that error configuration should be applied
+	 * @param group binding group
+	 * @param consumerProperties
+	 */
 	void register(Binding<T> binding, String group, C consumerProperties);
+
+	/**
+	 * Cleans any infrastructure created on the register method. It's up to an implementation to keep
+	 * track of any resources created so it can destroy it later.
+	 *
+	 * @param binding target binding that error configuration should be applied
+	 * @param group binding group
+	 * @param consumerProperties
+	 */
 	void destroy(Binding<T> binding, String group, C consumerProperties);
+
+	/**
+	 * Concrete implementations should override this and configure the binding target. Implementations should
+	 * keep track of what is being registered and then fetch the proper infrastructure resource to configure it.
+	 *
+	 * @param binding target binding that error configuration should be applied
+	 */
 	void configure(Binding<T> binding);
 }
