+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/error/LastSubscriberAwareChannel.java
@@ -14,17 +14,24 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.binder;
+package org.springframework.cloud.stream.error;
 
-import org.springframework.messaging.MessageHandler;
+import org.springframework.messaging.SubscribableChannel;
 
 /**
- * A marker interface designating a subscriber that must be the last.
+ * A channel that can ensure a {@code LastSubscriberMessageHandler} is always
+ * the last subscriber.
  *
  * @author Gary Russell
  * @since 1.3
  *
  */
-public interface LastSubscriberMessageHandler extends MessageHandler {
+interface LastSubscriberAwareChannel extends SubscribableChannel {
+
+	/**
+	 * Return the current subscribers.
+	 * @return the subscribers.
+	 */
+	int subscribers();
 
 }
